+++ src/ui/screens/topic_list.rs
@@ -89,11 +89,13 @@ pub fn render(frame: &mut Frame, app: &App, area: Rect) {
 
         render_selectable_list(
             frame,
+            app,
             main,
             &title,
             &items,
             Some(&["Name", "Partitions", "Replication", "Compression", "Messages"]),
             app.topic_list_selected_index,
+            true,
         );
     }